Refactor auth and event route modules
