+++ packages/server/src/routes/auth/constants.ts
@@ -1,0 +1,3 @@
+export const SYSTEM_TIMEZONE = "system";
+export const SYSTEM_DATE_TIME_LOCALE = "system";
+export const SYSTEM_THEME_PREFERENCE = "system";
